Fix undefined behavior for non-POD type T
Using the move-assignment operator on an uninitialized object (memory on which no constructor has been called) is undefined. The move assignment operator could be freeing resources in the object that is moved to, but that object doesn't exist in the raw memory. 
std::uninitialized_move() could be used, if the compiler supports C++17. Otherwise, we must force the use of the move constructor rather than the move assignment operator. We do this by using placement new.