Add default c'tor and clear function.
An empty map is a valid state, so there's no reason to not have a default c'tor. The std::map::clear function does exactly what a clear() on ArrayMap does, so there's no reason to hide the clear(). The replace function causes a compile error because being and end are ambiguous and could also be calls to operator() on the Iterators.